Add files via upload
Sudoku Game In Python is one of all time’s most famous puzzle games. The main objective of the Sudoku is to fill a 4 range chart (i.e. 4×4 grid) with numbers so that every segment of the row, column, and 2 range contains all the numbers between 1 and 4. This game is also a great brain game as a logic puzzle.